Don't force maximum volume on each bulk request; ignore archive entries with short length; null flush interval
We need to ignore archive entries with length < 4: no need to support older formats, but when we repack .tar files the directories (eg: index/, index/type) are considered part of this list.

When the flush interval is not null, the bulk client will asynchronously write entries to the database. We experienced record loss where anywhere between 1 and 999 records out of the bulk 1000 were not written, due to the separate flush calls. In this change, entries are only written in groups of (maxActionsPerBulkRequest) and none "get lost". The last block of entries may contain under this number, which is the only case where we want less than 1000 to be written by the bulk processor.

Changed flush interval to null in knapsack rather than support